test(openshell): consolidate manager fixtures
